Basics
Basic gameplay created. Create class gameplay with loop for game cycle. In there you can read key and move your player to right direction. Then you clear console and draw map again. Also create class Map with two dimensional array of ints and array of strings. It is not the best way to solve this problem, so I need to find out new one. So, now in game you can move to any possible direction and you can not go out of the range of the map.