Swap By Links
Add method that swaps two node by it is links. Although it is not working, there is no point in it. I write some code, that in theory should work, but in practice it is not, just to show my teacher, that I hava this code. Anyway, it is not in documantary, so I 100 percent right and can add this to my project, beacuse no one promise you, that it is going to work. Sorry. Love you, anyway.